requirements.txt
# YouTube API
google-api-python-client>=2.100.0
youtube-transcript-api>=0.6.1

# LLM APIs
anthropic>=0.18.0
openai>=1.12.0

# Telegram
python-telegram-bot>=21.0

# Database
aiosqlite>=0.19.0

# Scheduling
apscheduler>=3.10.4

# HTTP requests
httpx>=0.26.0
aiohttp>=3.9.0

# Configuration
pyyaml>=6.0.1
python-dotenv>=1.0.0

# Data validation
pydantic>=2.5.0
pydantic-settings>=2.1.0

# Utilities
structlog>=24.1.0
